refactor: 공통 코드 패키지를 도메인 구조로 이동
