+++ src/main/java/kr/kro/airbob/domain/commoncode/api/CommonCodeAdminController.java
@@ -1,4 +1,4 @@
-package kr.kro.airbob.common.code;
+package kr.kro.airbob.domain.commoncode.api;
 
 import java.util.List;
 import java.util.Locale;
@@ -15,27 +15,29 @@ import org.springframework.web.bind.annotation.RestController;
 
 import jakarta.validation.Valid;
 import kr.kro.airbob.common.dto.ApiResponse;
+import kr.kro.airbob.domain.commoncode.dto.CommonCodeAdminResponse;
+import kr.kro.airbob.domain.commoncode.dto.CommonCodeRequest;
+import kr.kro.airbob.domain.commoncode.service.CommonCodeAdminService;
 import lombok.RequiredArgsConstructor;
 
 /**
  * 공통 코드 관리 API (ADMIN 전용 — AdminAuthInterceptor 가 /api/v1/admin/** 보호).
- * 운영자가 라벨/정렬/노출여부를 바꾸거나 새 코드를 추가한다(배포 불필요).
  */
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/api/v1/admin/common-codes")
+@RequestMapping("/api")
 public class CommonCodeAdminController {
 
 	private final CommonCodeAdminService commonCodeAdminService;
 
 	// 비활성 포함 전체 조회(관리 화면용)
-	@GetMapping("/{group}")
+	@GetMapping("/v1/admin/common-codes/{group}")
 	public ResponseEntity<ApiResponse<List<CommonCodeAdminResponse>>> getAll(@PathVariable String group) {
 		List<CommonCodeAdminResponse> codes = commonCodeAdminService.getAll(group.toUpperCase(Locale.ROOT));
 		return ResponseEntity.ok(ApiResponse.success(codes));
 	}
 
-	@PostMapping("/{group}")
+	@PostMapping("/v1/admin/common-codes/{group}")
 	public ResponseEntity<ApiResponse<CommonCodeAdminResponse>> create(
 		@PathVariable String group,
 		@RequestBody @Valid CommonCodeRequest.Create request) {
@@ -44,7 +46,7 @@ public class CommonCodeAdminController {
 		return ResponseEntity.status(HttpStatus.CREATED).body(ApiResponse.success(created));
 	}
 
-	@PatchMapping("/{group}/{code}")
+	@PatchMapping("/v1/admin/common-codes/{group}/{code}")
 	public ResponseEntity<ApiResponse<CommonCodeAdminResponse>> update(
 		@PathVariable String group,
 		@PathVariable String code,
