+++ src/main/java/kr/kro/airbob/domain/commoncode/api/CommonCodeController.java
@@ -1,4 +1,4 @@
-package kr.kro.airbob.common.code;
+package kr.kro.airbob.domain.commoncode.api;
 
 import java.util.List;
 import java.util.Locale;
@@ -10,11 +10,12 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import kr.kro.airbob.common.dto.ApiResponse;
+import kr.kro.airbob.domain.commoncode.dto.CommonCodeResponse;
+import kr.kro.airbob.domain.commoncode.service.CommonCodeService;
 import lombok.RequiredArgsConstructor;
 
 /**
- * 공통 코드 조회 API. 프론트 셀렉트 박스/목록 노출용.
- * 예: GET /api/v1/common-codes/AMENITY_TYPE
+ * 공통 코드 조회 API. 프론트 셀렉트 박스/목록 노출용
  */
 @RestController
 @RequiredArgsConstructor
