+++ src/main/java/kr/kro/airbob/domain/commoncode/dto/CommonCodeRequest.java
@@ -1,4 +1,4 @@
-package kr.kro.airbob.common.code;
+package kr.kro.airbob.domain.commoncode.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.AccessLevel;
