+++ src/main/java/kr/kro/airbob/domain/commoncode/repository/CommonCodeGroupRepository.java
@@ -1,6 +1,8 @@
-package kr.kro.airbob.common.code;
+package kr.kro.airbob.domain.commoncode.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import kr.kro.airbob.domain.commoncode.entity.CommonCodeGroup;
 
 public interface CommonCodeGroupRepository
 	extends JpaRepository<CommonCodeGroup, String> {
