+++ src/test/java/kr/kro/airbob/domain/commoncode/service/CommonCodeAdminServiceIntegrationTest.java
@@ -1,4 +1,4 @@
-package kr.kro.airbob.common.code;
+package kr.kro.airbob.domain.commoncode.service;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -19,9 +19,11 @@ import org.testcontainers.utility.DockerImageName;
 
 import co.elastic.clients.elasticsearch.ElasticsearchClient;
 import io.awspring.cloud.s3.S3Template;
-import kr.kro.airbob.common.code.exception.CommonCodeDuplicateException;
-import kr.kro.airbob.common.code.exception.CommonCodeGroupNotFoundException;
-import kr.kro.airbob.common.code.exception.CommonCodeNotFoundException;
+import kr.kro.airbob.domain.commoncode.common.CommonCodeGroups;
+import kr.kro.airbob.domain.commoncode.dto.CommonCodeRequest;
+import kr.kro.airbob.domain.commoncode.exception.CommonCodeDuplicateException;
+import kr.kro.airbob.domain.commoncode.exception.CommonCodeGroupNotFoundException;
+import kr.kro.airbob.domain.commoncode.exception.CommonCodeNotFoundException;
 import kr.kro.airbob.search.repository.AccommodationSearchRepository;
 
 /**
